Added login functionality
